Added reverse coordinate converter.

+++ teleporter-chess-tests/SquareModelTests.cs
@@ -9,7 +9,7 @@ public class SquareModelTests
     [InlineData(Column.A, Row._1, 1, 1)]
     [InlineData(Column.E, Row._5, 5, 5)]
     [InlineData(Column.H, Row._8, 8, 8)]
-    public void GetOneBasedCoordinatesReturnsCorrectCoordinates(Column column, Row row, ushort expectedColumn, ushort expectedRow)
+    public void GetOneBasedCoordinatesReturnsCorrectCoordinates(Column column, Row row, uint expectedColumn, uint expectedRow)
     {
         Square square = new(column, row);
 
@@ -22,7 +22,7 @@ public class SquareModelTests
     [InlineData(Column.A, Row._1, 0, 0)]
     [InlineData(Column.E, Row._5, 4, 4)]
     [InlineData(Column.H, Row._8, 7, 7)]
-    public void GetZeroBasedCoordinatesReturnsCorrectCoordinates(Column column, Row row, ushort expectedColumn, ushort expectedRow)
+    public void GetZeroBasedCoordinatesReturnsCorrectCoordinates(Column column, Row row, uint expectedColumn, uint expectedRow)
     {
         Square square = new(column, row);
